pagin, sorting, filtering worked, add responseEntity

+++ src/main/java/com/education/projects/cars/manager/carsmanager/service/DBCarServiceImpl.java
@@ -12,7 +12,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.Collection;
 
 /**
@@ -23,106 +22,8 @@ public class DBCarServiceImpl implements CarService {
 
     @Autowired
     private CarRepository carRepository;
-    //private final CarRepository carRepository;
+    @Autowired
     private CarCriteriaRepository carCriteriaRepository;
-
-    /**
-     * Gets a car objects list, according to the current result set of the database,
-     * returns the list Car objects,
-     * parameter toPrint needs to choose whether the list can be printed
-     *
-     * @param toPrint is parameter to choose whether the list can be printed
-     * @return the list Car objects
-     */
-    private ArrayList<Car> getListCar(boolean toPrint) {
-        /*ArrayList<Car> carArrayList = new ArrayList<>();
-        try {
-            while (carRepository.getRs().next()) {
-                Car car = new Car();
-                setFieldCar(car);
-                carArrayList.add(car);
-            }
-            if(toPrint) printCarList(carArrayList);     //print result of query
-            System.out.println();
-            return carArrayList;
-        } catch (SQLException e) {
-            System.out.println(e.getMessage());
-        }*/
-        return null;
-    }
-
-    @Override
-    /**
-     * Gets the list of max cost Car objects from database, parameter list is null
-     * @param list is null
-     */
-    public ArrayList<Car> getMaxCostCar(ArrayList<Car> list) {
-        /*dbPoolService.statementExeQuery("SELECT * FROM Garage " +
-        "where cost = (select Max(Cost) from Garage);");*/
-        return getListCar(true);
-    }
-
-    @Override
-    /**
-     * Gets the list of min cost Car objects from database, parameter list is null
-     * @param list is null
-     */
-    public ArrayList<Car> getMinCostCar(ArrayList<Car> list) {
-   /*     dbPoolService.statementExeQuery("SELECT * FROM Garage " +
-                "where cost = (select Min(Cost) from Garage);");*/
-        return getListCar(true);
-    }
-
-    @Override
-    /**
-     * Finds the list of Car objects of the same brand from database, parameter list is null
-     * @param list is null
-     */
-    public ArrayList<Car> findBrandList(String searchBrand, ArrayList<Car> list) {
-        //dbPoolService.statementExeQuery("SELECT * FROM Garage WHERE Brand = '" + searchBrand +"';");
-        return getListCar(true);
-    }
-
-    @Override
-    /**
-     * Finds the list of Car objects of the same model from database, parameter list is null
-     * @param list is null
-     */
-    public ArrayList<Car> findModelList(String searchModel, ArrayList<Car> list) {
-        //dbPoolService.statementExeQuery("SELECT * FROM Garage WHERE Model = '" + searchModel +"';");
-        return getListCar(true);
-    }
-
-    @Override
-    /**
-     * Gets the list of Car objects ranked by price from database, parameter list is null
-     * @param list is null
-     */
-    public ArrayList<Car> getListByPriceRange(int startPrice, int endPrice, ArrayList<Car> list) {
-        /*dbPoolService.statementExeQuery("SELECT * FROM Garage WHERE Cost BETWEEN "
-                + startPrice + " AND " + endPrice +";");*/
-        return getListCar(true);
-    }
-
-    @Override
-    /**
-     * Get the list of Car objects sorted by price from database, parameter list is null
-     * @param list is null
-     */
-    public ArrayList<Car> sortListByPrice(ArrayList<Car> list) {
-        //dbPoolService.statementExeQuery("SELECT * FROM Garage ORDER BY Cost;");
-        return getListCar(true);
-    }
-
-    @Override
-    /**
-     * Gets the list of Car objects sorted by brand from database, parameter list is null
-     * @param list is null
-     */
-    public ArrayList<Car> sortListByBrand(ArrayList<Car> list) {
-        //dbPoolService.statementExeQuery("SELECT * FROM Garage ORDER BY Brand;");
-        return getListCar(true);
-    }
 
     /**
      * Creates a new Car object information in the database, returns a Car object from database by id
@@ -143,10 +44,15 @@ public class DBCarServiceImpl implements CarService {
      * @return Car object information from database by id
      * @throws Exception
      */
-    public Car updateAuto(Car car, Integer id) throws Exception {
-        Car carToChange = carRepository.getReferenceById(id);
-        car.setId(carToChange.getId());
-        return carRepository.save(car);
+    public Car updateAuto(Car car, Integer id) throws SQLException {
+        if(carRepository.existsById(id)) {
+            Car carToChange = carRepository.getReferenceById(id);
+            car.setId(carToChange.getId());
+            return carRepository.save(car);
+        } else {
+            throw new SQLException("The car with id = " + id +" wasn't found");
+        }
+
     }
 
     /**
@@ -166,7 +72,10 @@ public class DBCarServiceImpl implements CarService {
      * @throws SQLException
      */
     public Car getCarById(Integer id) throws SQLException {
-        return carRepository.getReferenceById(id);
+        if(carRepository.existsById(id))
+            return carRepository.getReferenceById(id);
+        else
+            throw new SQLException("The car with id = " + id +" wasn't found");
     }
 
     /**
@@ -174,8 +83,10 @@ public class DBCarServiceImpl implements CarService {
      *
      * @param id is a row in database
      */
-    public void deleteCarById(Integer id) {
-        carRepository.deleteById(id);
+    public void deleteCarById(Integer id) throws SQLException {
+        if(carRepository.existsById(id))
+            carRepository.deleteById(id);
+        else throw new SQLException("The car with id = " + id +" wasn't found");
     }
 
     /**
